Update the tests to create/tear-down dev-boxes
